+++ packages/sitecore-jss-nextjs/src/context/index.ts
@@ -1,1 +1,0 @@
-export { Context, ContextConfig, SDK } from './context';